Nothing to build (#7)
* Nothing to build

* Name the container based on input

+++ test.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
 # Assuming FFMPEG_DOCKER set
-IMAGE=${IMAGE:-ripfile}
 
 set -ex
-curl -L -o test.mpg https://alcorn.com/wp-content/downloads/test-files/AC3AlcornTest_HD.mpg
-HWACCEL=n INPUT="test.mpg" TITLE="Test" YEAR="2019" LOG=n ./ripFile.sh 
+if [[ ! -f test.mpg ]]; then
+	curl -L -o test.mpg https://alcorn.com/wp-content/downloads/test-files/AC3AlcornTest_HD.mpg
+fi
+HWACCEL=n INPUT=test.mpg TITLE="Test Title" YEAR=2019 DOCKER_DAEMON=n ./ripFile.sh 
+docker rm test.mpg
